requirements.txt
# Core dependencies
numpy>=2.0.0
torch>=2.0.0
torchaudio>=2.0.0
openai-whisper>=20240930
pydub>=0.25.0
soundfile>=0.12.0

# Google Cloud Translation API
google-cloud-translate>=3.0.0

# NLP and text processing
sentence-transformers>=4.0.0
scikit-learn>=1.0.0
regex>=2024.0.0

# Utilities
tqdm>=4.0.0
requests>=2.0.0
PyYAML>=6.0.0
